test(react-native): stop the message content runtime tests from timing out on CI
the two runtime cases lived in `MessageContent.test.tsx`, which mocks `@assistant-ui/store` for the dispatch tests, so they had to undo that per test with `vi.doUnmock` plus `vi.resetModules()` and then `await import("@assistant-ui/core/react")` inside the test body. re-evaluating the whole core react module graph costs 568ms on an idle mac and about 720ms in a linux container, and a loaded CI runner crosses the 5000ms test timeout. the timed-out case never reaches its `finally` unmount, so the leaked act queue makes the sibling case read the pre-update DOM and fail with `expected 'running' to be 'complete'`.

the cases never needed the mock, so they move to `MessageContent.runtime.test.tsx` with plain static imports. the assertions are unchanged and the tool-call case drops to 18ms on mac and 22ms in the same container.
