+++ packages/react-native/src/primitives/message/MessageContent.test.tsx
@@ -1,7 +1,6 @@
 import { act, type ReactElement } from "react";
 import { createRoot, type Root } from "react-dom/client";
 import { afterEach, beforeEach, describe, expect, it, vi } from "vitest";
-import type { MessagePartState, ThreadMessageLike } from "@assistant-ui/core";
 import { MessageContent } from "./MessageContent";
 
 type AnyPart = { type: string; [key: string]: unknown };
@@ -275,84 +274,4 @@ describe("MessageContent", () => {
 
     expect(container.textContent).toBe("A[tool][data]");
   });
-});
-
-describe("MessageContent with a runtime", () => {
-  it.each(["tool-call", "data"] as const)(
-    "renders derived status for registered %s UIs",
-    async (type) => {
-      vi.doUnmock("@assistant-ui/store");
-      vi.resetModules();
-      // Static imports retain the store mock used by the dispatch tests.
-      const { MessageContent: RuntimeContent } =
-        await import("./MessageContent");
-      const {
-        AssistantRuntimeProvider,
-        MessageByIndexProvider,
-        useExternalStoreRuntime,
-        useAssistantToolUI,
-        useAssistantDataUI,
-      } = await import("@assistant-ui/core/react");
-
-      const Status = ({ status }: Pick<MessagePartState, "status">) => (
-        <span>{status.type}</span>
-      );
-      const Registrations = () => {
-        useAssistantToolUI({ toolName: "search", render: Status });
-        useAssistantDataUI({ name: "chart", render: Status });
-        return <RuntimeContent />;
-      };
-      const App = ({ message }: { message: ThreadMessageLike }) => {
-        const runtime = useExternalStoreRuntime({
-          messages: [message],
-          convertMessage: (value) => value,
-          onNew: async () => {},
-        });
-        return (
-          <AssistantRuntimeProvider runtime={runtime}>
-            <MessageByIndexProvider index={0}>
-              <Registrations />
-            </MessageByIndexProvider>
-          </AssistantRuntimeProvider>
-        );
-      };
-      const container = document.createElement("div");
-      const root = createRoot(container);
-      try {
-        for (const status of [
-          "running",
-          "requires-action",
-          "complete",
-        ] as const) {
-          if (type === "data" && status === "requires-action") continue;
-          const message: ThreadMessageLike = {
-            id: "m1",
-            role: "assistant",
-            status:
-              status === "running"
-                ? { type: status }
-                : status === "complete"
-                  ? { type: status, reason: "stop" }
-                  : { type: status, reason: "tool-calls" },
-            content: [
-              type === "data"
-                ? { type, name: "chart", data: { value: 1 } }
-                : {
-                    type,
-                    toolName: "search",
-                    toolCallId: "c1",
-                    args: {},
-                    argsText: "{}",
-                    ...(status === "complete" ? { result: "done" } : {}),
-                  },
-            ],
-          };
-          await act(async () => root.render(<App message={message} />));
-          expect(container.textContent).toBe(status);
-        }
-      } finally {
-        await act(async () => root.unmount());
-      }
-    },
-  );
 });
